requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0

# Geospatial & Remote Sensing
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Sentinel-1 Data Retrieval
asf-search>=6.0.0

# InSAR Processing
# pygmtsar  # GMTSAR Python wrapper (설치 가이드 별도 참조)
# isce2     # NASA JPL ISCE2 (설치 가이드 별도 참조)

# Visualization
folium>=0.14.0
plotly>=5.14.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
rich>=13.0.0

# Development & Testing
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
